dograf/requirements.txt
psutil==5.7.2
psycopg2-binary==2.8.6
